requirements.txt
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.1
Django==3.2.20
django-cors-headers==3.9.0
django-extra-fields==3.0.2
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0
djoser==2.2.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
pytz==2020.1
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.3.1
django_cron